+++ apiserver/internal/model/site.go
@@ -3,7 +3,7 @@ package model
 type SiteModel struct {
 	Model
 	Domain    string `json:"domain" gorm:"uniqueIndex;type:varchar(255);not null" binding:"required,fqdn"`
-	Status    int16  `json:"status" gorm:"type:smallint;not null;default:0"`
+	Status    int8   `json:"status" gorm:"type:smallint;not null;default:0"`
 	Remark    string `json:"remark" gorm:"type:text"`
 	DeletedAt int64  `json:"-" gorm:"not null;default:0"`
 }
@@ -21,7 +21,7 @@ type SiteReq struct {
 type SiteRsp struct {
 	Model
 	Domain  string           `json:"domain"`
-	Status  int16            `json:"status"`
+	Status  int8             `json:"status"`
 	Remark  string           `json:"remark"`
 	Config  *SiteConfigRsp   `json:"config"`
 	Origins []*SiteOriginRsp `json:"origins"`
