+++ apiserver/internal/model/whitelist_ip.go
@@ -4,7 +4,7 @@ type WhitelistIPModel struct {
 	Model
 	IP     string `json:"ip" gorm:"uniqueIndex;type:varchar(39)" binding:"required,ip4_addr|cidrv4"`
 	Remark string `json:"remark" gorm:"type:text"`
-	Status int16  `json:"status" gorm:"type:smallint;not null;default 1"`
+	Status int8   `json:"status" gorm:"type:smallint;not null;default 1"`
 }
 
 func (WhitelistIPModel) TableName() string {
